requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
python-dotenv
gunicorn